backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "kagglehub>=0.3.13",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
